finish reorg (#371)
