--- sramgen/src/layout/v2/mod.rs
@@ -1,2 +1,0 @@
-pub mod array;
-pub mod precharge;
