+++ sramgen/src/v2/bitcell_array/schematic.rs
@@ -1,0 +1,10 @@
+use super::SpCellArray;
+
+impl SpCellArray {
+    pub(crate) fn schematic(
+        &self,
+        _ctx: &mut substrate::schematic::context::SchematicCtx,
+    ) -> substrate::error::Result<()> {
+        Ok(())
+    }
+}
