+++ sramgen/src/v2/precharge/schematic.rs
@@ -1,0 +1,10 @@
+use super::Precharge;
+
+impl Precharge {
+    pub(crate) fn schematic(
+        &self,
+        _ctx: &mut substrate::schematic::context::SchematicCtx,
+    ) -> substrate::error::Result<()> {
+        Ok(())
+    }
+}
